Fixed Illegal string offset due to (un)archiving channels
The payload that is sent over from slack when you (un)archive a
channel collapses the "channel" sub-array to just be a string
containing the ID of the channel. That resulted in an
Illegal string offset error being thrown if the bot was in a channel
that was archived or unarchived.